+++ zscaler/zpa/app_segments_ba.py
@@ -19,7 +19,7 @@ from zscaler.api_client import APIClient
 from zscaler.request_executor import RequestExecutor
 from zscaler.zpa.models.application_segment import ApplicationSegments
 from zscaler.zpa.app_segment_by_type import ApplicationSegmentByTypeAPI
-from zscaler.utils import format_url, add_id_groups
+from zscaler.utils import format_url, transform_common_id_fields
 from zscaler.types import APIResult
 import logging
 
@@ -294,7 +294,7 @@ class ApplicationSegmentBAAPI(APIClient):
         if "clientless_app_ids" in body:
             body["clientlessApps"] = body.pop("clientless_app_ids")
 
-        add_id_groups(self.reformat_params, kwargs, body)
+        transform_common_id_fields(self.reformat_params, kwargs, body, coerce_ids=False)
 
         request, error = self._request_executor.create_request(http_method, api_url, body=body, params=params)
         if error:
@@ -461,7 +461,7 @@ class ApplicationSegmentBAAPI(APIClient):
         else:
             body["udpPortRange"] = []  # Explicitly clear if not provided
 
-        add_id_groups(self.reformat_params, kwargs, body)
+        transform_common_id_fields(self.reformat_params, kwargs, body, coerce_ids=False)
 
         request, error = self._request_executor.create_request(http_method, api_url, body, {}, params)
         if error: